backend/requirements.txt
fastapi 
uvicorn 
openai 
python-dotenv
